Implemented new approach for handling failed jobs (#112)
* Implemented new approach for handling failed jobs

* Moved comment

* Rephrased comment

* Updating FirstOrDefault linq expression for existing failed job

* Rename collections

* Tested and fixed a few minor bugs

* Get list instead of single on timeouts there is a risk that multiple failed jobs gets added

* Added cosmetic changes to failed job view

* Ensuring that Invalidation process inherits the same behavoir

SaveOrUPdateFailed jobs on the jobs handler has been has been made public, so this can be called from outside the implementation. The approach is very much the same as HandleJobs and HandlePendingJobs methods.

* Changed error message for timed out jobs

* Ensuring that processed jobs gets deleted

Jobs that has been set to the state of processing is now being deleted.

* Get failed jobs by both entityId and culture

When we remove failed jobs we do it both by entity id and culture, we should do the same thing when fetching failed jobs.

* Changing the order in which SaveOrUpdateFailedJobs is called

We want to delete the queued jobs before we call SaveOrUpdatedFailedJobs, since this SaveOrUpdateFailedJobs throws an exception. We need to complete all actions, before throwing the exception

---------

Co-authored-by: jesperweber <jweber@jweber.dk>